+++ src/dve/core_engine/backends/implementations/duckdb/readers/xml.py
@@ -1,19 +1,16 @@
 # mypy: disable-error-code="attr-defined"
 """An xml reader to create duckdb relations"""
 
-from typing import Dict, Type
+from typing import Dict, Optional, Type
 
 import polars as pl
-from duckdb import DuckDBPyConnection, DuckDBPyRelation
+from duckdb import DuckDBPyConnection, DuckDBPyRelation, default_connection
 from pydantic import BaseModel
 
 from dve.core_engine.backends.base.reader import read_function
-from dve.core_engine.backends.implementations.duckdb.duckdb_helpers import (
-    duckdb_write_parquet,
-    get_polars_type_from_annotation,
-)
+from dve.core_engine.backends.implementations.duckdb.duckdb_helpers import duckdb_write_parquet
 from dve.core_engine.backends.readers.xml import XMLStreamReader
-from dve.core_engine.backends.utilities import stringify_model
+from dve.core_engine.backends.utilities import get_polars_type_from_annotation, stringify_model
 from dve.core_engine.type_hints import URI
 
 
@@ -21,15 +18,15 @@ from dve.core_engine.type_hints import URI
 class DuckDBXMLStreamReader(XMLStreamReader):
     """A reader for XML files"""
 
-    def __init__(self, ddb_connection: DuckDBPyConnection, **kwargs):
-        self.ddb_connection = ddb_connection
+    def __init__(self, ddb_connection: Optional[DuckDBPyConnection] = None, **kwargs):
+        self.ddb_connection = ddb_connection if ddb_connection else default_connection
         super().__init__(**kwargs)
 
     @read_function(DuckDBPyRelation)
     def read_to_relation(self, resource: URI, entity_name: str, schema: Type[BaseModel]):
         """Returns a relation object from the source xml"""
         polars_schema: Dict[str, pl.DataType] = {  # type: ignore
-            fld.name: get_polars_type_from_annotation(fld.type_)
+            fld.name: get_polars_type_from_annotation(fld.annotation)
             for fld in stringify_model(schema).__fields__.values()
         }