+++ src/dve/core_engine/backends/implementations/spark/spark_helpers.py
@@ -16,6 +16,8 @@ from typing import (
     Callable,
     ClassVar,
     Dict,
+    Generator,
+    Iterator,
     List,
     Optional,
     Set,
@@ -345,18 +347,19 @@ def _spark_read_parquet(self, path: URI, **kwargs) -> DataFrame:
 
 
 def _spark_write_parquet(  # pylint: disable=unused-argument
-    self, entity: DataFrame, target_location: URI, **kwargs
+    self, entity: Union[Iterator[Dict[str, Any]], DataFrame], target_location: URI, **kwargs
 ) -> URI:
     """Method to write parquet files from type cast entities
     following data contract application
     """
-    _options = {"schema": entity.schema, **kwargs}
-    (
-        entity.write.options(**_options)  # type: ignore
-        .format("parquet")
-        .mode("overwrite")
-        .save(target_location)
-    )
+    _options: Dict[str, Any] = {**kwargs}
+    if isinstance(entity, Generator):
+        _writer = self.spark_session.createDataFrame(entity).write
+    else:
+        _options["schema"] = entity.schema  # type: ignore
+        _writer = entity.write  # type: ignore
+
+    (_writer.options(**_options).format("parquet").mode("overwrite").save(target_location))
     return target_location