+++ src/dve/pipeline/duckdb_pipeline.py
@@ -2,13 +2,14 @@
 
 from typing import Optional, Type
 
-from duckdb import DuckDBPyConnection
+from duckdb import DuckDBPyConnection, DuckDBPyRelation
 
 from dve.core_engine.backends.base.reference_data import BaseRefDataLoader
 from dve.core_engine.backends.implementations.duckdb.auditing import DDBAuditingManager
 from dve.core_engine.backends.implementations.duckdb.contract import DuckDBDataContract
 from dve.core_engine.backends.implementations.duckdb.duckdb_helpers import duckdb_get_entity_count
-from dve.core_engine.backends.implementations.duckdb.rules import DuckDBStepImplemetations
+from dve.core_engine.backends.implementations.duckdb.rules import DuckDBStepImplementations
+from dve.core_engine.models import SubmissionInfo
 from dve.core_engine.type_hints import URI
 from dve.pipeline.pipeline import BaseDVEPipeline
 
@@ -35,9 +36,17 @@ class DDBDVEPipeline(BaseDVEPipeline):
             audit_tables,
             job_run_id,
             DuckDBDataContract(connection=self._connection),
-            DuckDBStepImplemetations.register_udfs(connection=self._connection),
+            DuckDBStepImplementations.register_udfs(connection=self._connection),
             rules_path,
             processed_files_path,
             submitted_files_path,
             reference_data_loader,
+        )
+
+    # pylint: disable=arguments-differ
+    def write_file_to_parquet(  # type: ignore
+        self, submission_file_uri: URI, submission_info: SubmissionInfo, output: URI
+    ):
+        return super().write_file_to_parquet(
+            submission_file_uri, submission_info, output, DuckDBPyRelation
         )